# docs/api.yaml
openapi: 3.0.3
info:
  title: BeLib Implemented API
  version: 0.6.0
  description: |
    BeLibで現在実装されている /api/v1 HTTP contract を記録する。
    将来の正式要件や未実装endpointは docs/requirements.md と docs/mvp_plan.md を参照する。
    認証はBetter AuthのCookie sessionを使用し、Bearer/JWTをこの文書の契約として扱わない。
    Better Auth自身の /api/auth/* はライブラリ管理のため、このversioned API定義には複製しない。
servers:
  - url: http://localhost:3000/api/v1
    description: Local backend

paths:
  /me:
    get:
      summary: 認証中のユーザーを取得
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /books:
    get:
      summary: 閲覧権限のある書籍一覧を取得
      description: |
        RoleBookPermissionで現在ユーザーのroleに許可された、論理削除されていない書籍だけを返す。
        現実装ではpage/limitのみbackend queryとして有効。検索、category、readStatus filterは未実装。
        一覧のreadStatusは各書籍についてrequesting user自身のReadingInfoだけを参照し、未作成時はunreadを返す。
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Authorized books
          content:
            application/json:
              schema:
                type: object
                required: [total, page, limit, books]
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  /books/{bookId}:
    parameters:
      - $ref: '#/components/parameters/BookId'
    get:
      summary: 閲覧権限のある書籍詳細を取得
      responses:
        '200':
          description: Book detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /books/{bookId}/file:
    parameters:
      - $ref: '#/components/parameters/BookId'
    get:
      summary: 認可済みPDFを取得
      description: |
        現在の実装はPDF専用。認証・role permission確認後、storage adapter経由でbackendからprivate streamする。
        ブラウザへpublic URLやpresigned URLは返さない。local adapterまたは設定済みR2 adapterを使用する。
      responses:
        '200':
          description: PDF stream
          headers:
            Cache-Control:
              schema:
                type: string
              example: private, no-store
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/FileNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /books/{bookId}/reading-info:
    parameters:
      - $ref: '#/components/parameters/BookId'
    get:
      summary: 現在ユーザーの読書情報を取得
      responses:
        '200':
          description: Reading info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookNotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      summary: PDFの現在ページを保存
      description: |
        currentPageとPDFのtotalPagesを受け取り、ReadingInfoをupsertする。
        最終ページ到達時はbackend側でcompletedへ遷移し、completed後の保存ではcompletedを維持する。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadingInfoUpdateRequest'
      responses:
        '200':
          description: Saved reading info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadingInfo'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /admin/categories:
    get:
      summary: 管理者が有効なカテゴリ一覧を取得
      description: |
        管理者だけが取得できる。isActive=trueのカテゴリをdisplayOrder、idの昇順で返す。
      responses:
        '200':
          description: Active categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCategoriesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /admin/books:
    post:
      summary: 管理者が書籍メタデータを登録
      description: |
        title、active categoryIdと、現在のBook schemaで表現できるメタデータを受け取る。
        ファイルはこのendpointでは登録しない。
        現実装はuser role向けRoleBookPermissionも作成するため、正式な公開範囲選択要件とはまだ一致していない。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminBookCreateRequest'
      responses:
        '201':
          description: Created book metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBookCreateResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  /admin/books/{bookId}/files:
    parameters:
      - $ref: '#/components/parameters/BookId'
    post:
      summary: 管理者がPDFを登録
      description: |
        現在はPDFのみ、最大200MB。metadata、PDF header、SHA-256重複を検証し、storage adapterへの保存成功後にBookFileを作成する。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created BookFile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookFile'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/BookNotFound'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalError'

components:
  parameters:
    BookId:
      name: bookId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1

  responses:
    InvalidRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Permission denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BookNotFound:
      description: Book not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    FileNotFound:
      description: PDF file not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true

  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string

    CurrentUser:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string

    CategorySummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string

    AdminCategoriesResponse:
      type: object
      required: [categories]
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummary'

    Book:
      type: object
      required:
        - id
        - title
        - category
        - pageTurnDirection
        - readStatus
        - hasFile
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        authorName:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        publisher:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/CategorySummary'
        pageTurnDirection:
          type: string
          enum: [ltr, rtl]
        readStatus:
          type: string
          enum: [unread, reading, completed]
        hasFile:
          type: boolean
        fileType:
          type: string
          enum: [pdf, epub]
          nullable: true
        fileSize:
          type: integer
          nullable: true
        originalFileName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ReadingInfo:
      type: object
      required: [bookId, currentPage, readStatus]
      properties:
        bookId:
          type: integer
        currentPage:
          type: integer
          minimum: 1
        readStatus:
          type: string
          enum: [unread, reading, completed]

    ReadingInfoUpdateRequest:
      type: object
      required: [currentPage, totalPages]
      additionalProperties: false
      properties:
        currentPage:
          type: integer
          minimum: 1
        totalPages:
          type: integer
          minimum: 1

    AdminBookCreateRequest:
      type: object
      required: [title, categoryId]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
        categoryId:
          type: integer
          minimum: 1
        authorName:
          type: string
          nullable: true
          maxLength: 255
        publisher:
          type: string
          nullable: true
          maxLength: 255
        publishedAt:
          type: string
          format: date
          nullable: true
        pageTurnDirection:
          type: string
          enum: [ltr, rtl]
          default: ltr
        description:
          type: string
          nullable: true

    AdminBookCreateResponse:
      allOf:
        - $ref: '#/components/schemas/AdminBookRecord'

    AdminBookRecord:
      type: object
      required: [id, title, categoryId, category, pageTurnDirection]
      properties:
        id:
          type: integer
        title:
          type: string
        authorName:
          type: string
          nullable: true
        publisher:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        categoryId:
          type: integer
        category:
          $ref: '#/components/schemas/CategorySummary'
        pageTurnDirection:
          type: string
          enum: [ltr, rtl]
        description:
          type: string
          nullable: true

    BookFile:
      type: object
      required:
        - id
        - extension
        - mimeType
        - fileUrl
        - originalFileName
        - storedFileName
        - fileSize
        - fileHash
        - bookId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        extension:
          type: string
          enum: [pdf]
        mimeType:
          type: string
          enum: [application/pdf]
        fileUrl:
          type: string
          description: Opaque protected storage key or locator. It is never a public URL.
        originalFileName:
          type: string
        storedFileName:
          type: string
        fileSize:
          type: integer
        fileHash:
          type: string
        bookId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time